backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
psycopg2-binary==2.9.9
pydantic==2.6.0
python-multipart
python-jose[cryptography]
passlib[bcrypt]
pandas
openpyxl
pdfplumber
mt940
python-dotenv
google-generativeai
websockets
xlsxwriter
